fix: convert components high on tree to server component per nextjs recommendations
