+++ src/app/components/toc.js
@@ -1,14 +1,29 @@
-export function ToC({ headings }) {
-    return (
-        // <div className="w-1/4 p-4 sticky top-0">
-        <div className="p-1 w-80 sticky top-0 border-r-2">
-            <ul className="list-none">
-                {headings.map((heading) => (
-                    <li key={heading.id} className="">
-                        <a className="no-underline text-black text-xs" href={`#${heading.id}`}>{heading.title.substring(0,)}</a>
-                    </li>
-                ))}
-            </ul>
-        </div>
-    )
+"use client";
+import { useEffect, useState } from "react"
+
+export function ToC() {
+
+  const [headings, setHeadings] = useState([])
+
+  useEffect(() => {
+    // Get all the headings on the page
+    const headingElements = document.querySelectorAll("h1, h2")
+    const headings = Array.from(headingElements).map((heading) => ({
+      title: heading.innerText,
+      id: heading.innerText.toLowerCase().replace(/\s/g, "-"),
+    }))
+    setHeadings(headings)
+  }, [])
+
+  return (
+    <div className="p-1 w-80 sticky top-0 border-r-2">
+      <ul className="list-none">
+        {headings.map((heading) => (
+          <li key={heading.id} className="">
+            <a className="no-underline text-black text-xs" href={`#${heading.id}`}>{heading.title.substring(0,)}</a>
+          </li>
+        ))}
+      </ul>
+    </div>
+  )
 }
